Inlined OverviewModelImpl
